requirements.txt
flask==2.0.1
flask-pymongo==2.3.0
flask-jwt-extended==4.3.1
bcrypt==0.7.1
flask-cors==3.0.10
python-dotenv==0.19.1
marshmallow==3.13.0
pymongo[srv]==3.12.0
bcrypt==3.2.0
